--- crates/egui_kittest/tests/snapshots/atom_layout_nesting.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e448998e5d18f3d5704e1280794ad5732087afab73fb591ad8025810a09f2db0
-size 6338
